push

+++ utils/api.tsx
@@ -39,12 +39,12 @@ export const addTemplate = async (teks: string, authKey: string): Promise<AxiosR
 
 // Delete template function
 export const deleteTemplate = async (id: number, authKey: string): Promise<AxiosResponse> => {
-  return api.delete(`/templates?id=${id}`, {
+  return api.delete(`/admin/templates?id=${id}`, {
     headers: { "x-auth-key": authKey },
   });
 };
 
 // Confirm template function
 export const confirmTemplate = async (id: number, authKey: string): Promise<AxiosResponse> => {
-  return api.put(`/templates?id=${id}`, {}, { headers: { "x-auth-key": authKey } });
+  return api.put(`/admin/templates?id=${id}`, {}, { headers: { "x-auth-key": authKey } });
 };
